+++ Azure/Controllers/RolController.cs
@@ -35,6 +35,11 @@ namespace Azure.Controllers
             if(!_tokenService.ValidateToken(token))
                 return new UnauthorizedObjectResult("Gecersiz Token");
 
+            var claims = _tokenService.GetTokenClaims(token);
+            
+            if(claims["role"] != "Admin")
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
+
             try
             {
                 var roller = await _rollerRepo.GetRollerAsync();
@@ -58,6 +63,11 @@ namespace Azure.Controllers
 
             if(!_tokenService.ValidateToken(token))
                 return new UnauthorizedObjectResult("Gecersiz Token");
+
+            var claims = _tokenService.GetTokenClaims(token);
+            
+            if(claims["role"] != "Admin")
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
 
             try
             {
@@ -83,6 +93,11 @@ namespace Azure.Controllers
             if(!_tokenService.ValidateToken(token))
                 return new UnauthorizedObjectResult("Gecersiz Token");
 
+            var claims = _tokenService.GetTokenClaims(token);
+
+            if(claims["role"] != "Admin")
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
+
             try
             {
                 var rol = await _rollerRepo.DeleteRolByIdAsync(id);
@@ -106,6 +121,12 @@ namespace Azure.Controllers
 
             if(!_tokenService.ValidateToken(token))
                 return new UnauthorizedObjectResult("Gecersiz Token");
+
+            var claims = _tokenService.GetTokenClaims(token);
+
+            if(claims["role"] != "Admin")
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
+            
             
             try
             {
@@ -134,6 +155,11 @@ namespace Azure.Controllers
 
             if(!_tokenService.ValidateToken(token))
                 return new UnauthorizedObjectResult("Gecersiz Token");
+
+            var claims = _tokenService.GetTokenClaims(token);
+            
+            if(claims["role"] != "Admin")
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
 
             try
             {
